tests/wpt/web-platform-tests/shadow-dom/styles/test-010.html

<!DOCTYPE html>
<!--
Distributed under both the W3C Test Suite License [1] and the W3C
3-clause BSD License [2]. To contribute to a W3C Test Suite, see the
policies and contribution forms [3].

[1] http://www.w3.org/Consortium/Legal/2008/04-testsuite-license
[2] http://www.w3.org/Consortium/Legal/2008/03-bsd-license
[3] http://www.w3.org/2004/10/27-testcases
-->
<html>
<head>
<title>Shadow DOM Test: A_06_00_11</title>
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="help" href="http://www.w3.org/TR/2013/WD-shadow-dom-20130514/#styles">
<meta name="assert" content="Styles:the styles of the shadow insertion point node are inherited by the child nodes of the shadow root of the shadow tree, distributed to this shadow insertion point">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script>
// Test fails. See https://bugs.webkit.org/show_bug.cgi?id=103625
test(unit(function (ctx) {
	var d = newRenderedHTMLDocument(ctx);

    var host = d.createElement('div');
    d.body.appendChild(host);

	//Old tree
	var s1 = host.createShadowRoot();

	var div1 = d.createElement('div');
	div1.setAttribute('style', 'font-size: 10px');
	div1.innerHTML = '<span id="shd1">This is an old shadow tree</span>';
	s1.appendChild(div1);

	var height1 = s1.querySelector('#shd1').offsetHeight;

	assert_true(height1 > 0, 'Element height should be greater than zero');

	//younger tree
	var s2 = host.createShadowRoot();
	var div2 = d.createElement('div');
	div2.innerHTML =  '<shadow style="font-size:20px"></shadow>';
	s2.appendChild(div2);

	assert_true(s1.querySelector('#shd1').offsetHeight > height1,
		'Shadow insertion point style must be aplied to the child nodes of ' +
    	'the shadow host that are assigned to this insertion point');
}), 'A_06_00_11_T01');
</script>
</body>
</html>
